+++ data/socialIconsItems.js
@@ -1,0 +1,41 @@
+import Icon from '../icons';
+
+const socialIconsItems = [
+    {
+        id: 1,
+        href: 'https://www.snapchat.com/',
+        logoClassName: 'styles.socialIcon',
+        name: Icon.SNAPCHAT,
+        sizes: { height: 40, width: 'auto' },
+    },
+    {
+        id: 2,
+        href: 'https://www.facebook.com/',
+        logoClassName: 'styles.socialIcon',
+        name: Icon.FACEBOOK,
+        sizes: { height: 40, width: 'auto' },
+    },
+    {
+        id: 3,
+        href: 'https://www.linkedin.com/',
+        logoClassName: 'styles.socialIcon',
+        name: Icon.LINKEDIN,
+        sizes: { height: 40, width: 'auto' },
+    },
+    {
+        id: 4,
+        href: 'https://www.skype.com/',
+        logoClassName: 'styles.socialIcon',
+        name: Icon.SKYPE,
+        sizes: { height: 40, width: 'auto' },
+    },
+    {
+        id: 5,
+        href: 'https://twitter.com/',
+        logoClassName: 'styles.socialIcon',
+        name: Icon.TWITTER,
+        sizes: { height: 40, width: 'auto' },
+    },
+];
+
+export default socialIconsItems;
